Fix way rotation creating invalid objects
Due to floating point inaccuracies in extreme circumstances, that is
fully zoomed in etc. the cosinus of the rotation angle could come out
slightly larger than 1 which caused arccos to return a NaN. THis commit
clamps the relevant value to legal values and adds a few further
improvements to the rotating code.