Add facility to check domains/URL in website fields
This adds a facility to take a url or naked domain and test if it is
reachable by https or http, with a preference for https. If the domain
is reachable it will return a correctly formated URL.

Further this refactors DialogRow to be more generic so that it can be
used as the base for further "popup a dialog to edit" fields.