Re-factor OSC and Undo code
This re-factors the undo code to always use the in memory copy of an
element to avoid issues when an element object is replaced, which
happens when applying OSC files and merging data.

The OSC code now uses the similar logic as the mergeData method to
ensure that the operation can be cleanly aborted (fixes crashes when OSC
files couldn't completely be applied).