Support fixed length list/array semantics for multi-selects
This adds support for a "value_count_key" attribute in presets that
refers to a tag that holds the number of values that should be present.

MultiTextRow has been modified so that when such a tag is present it
will try to always show the correct number of rows and concatenate them
with array semantics.