Adjust GPX coordinate precision
Use 8 digits for latitude and longitude.  This is about 1 mm, which is
silly for internal GPS of today's phones, but not for dual-frequwency
RTK.

Use 3 digits for altitude, which is 1 mm.  That's even better than the
best current equipment.

These values are chosen on the theory that one should be able to see
GPS errors in the recorded data, that the data format should never be
the limiting factor, and that recorded precision should not be vastly
more than is necessary.