Try to persist file permissions and catch SecurtiyExceptions
GEOJSON files are not copied in to the app specific storage for a number
of reasons, however this means that the read permissions will be lost
over a reboot.

This commit tries to persist the permissions and if that isn't possible
display a toast. Further we now catch the relevant security exception
again with a toast and remove the layer if the file can't be read.