Add mechanism to indicate if non-OSM objects can be selected
This adds a mechanism for EasyEdit modes to indicate if non-OSM objects
should be selectable when they are active or not. Currently this is only
an issue in the multi-select "ExtendSelection" mode.

A longer term, more elegant solution, would be to allow the mode to
indicate if it allows selection just in a specific layer or not. Which
would be useful if there would be editing support for non-OSM elements.