Fix spurious relation backlinks
If undoing an action which included removing an relation, backlinks
would not be correctly recreated. This caused them to turn up in
disambiguation dialogs, but had no other effect than being annoying.

THis might have to be revisited as it involves touching all elements
currently in storage and may be slow.