Compute pseudo accuracy from NMEA
The standard NMEA sentences do not contain an estimate of accuracy,
and thus Vespucci does not show an accuracy circle when using NMEA.
When surveying with RTK, it's very important to know if the device is
in normal mode vs RTK-float vs RTK-fixed.

Therefore, extract the fix type from GGA, and estimate accuracy,
scaled by HDOP, of 0.1 for RTK/fixed 1m for RTK/float, and 5m
(differential) or 10m (autonomous).