Don't set map location if showGps isn't set
This checks the show GPS preference in gotoCurrentLocation and only sets
the map location (which triggers display of the arrow) if it is set.

Further gotoCurrentLocation now actually uses the location manager last
location if the tracker service is present, but doesn't have a last
location.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/1588