Split the Preset class (4400 loc) in to individual classes
The Preset class used a hierarchical subclass design for the the
elements which led to the class becoming very large and the internal
dependencies being hard to follow. This commit splits the subclasses in
to their own files as a first step in a larger refactoring of the preset
system.