This moves the alternative preset display to above the MRU presets
The alternative preset display was hidden behind a menu, this moves it
to the top of the MRU presets.

Further when applying presets we try to remove fixed keys of
alternatives of the preset we are applying and have slightly improved
the reporting of overwritten/removed tags.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/1522