Remove search results when search term is too short
This changes the behaviour of the preset search to remove the result
fragment if the search term is shorter than chars (currently hardwired).

Further it fixes the race condition that now and then caused two or more
result fragments to be added.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/1793