Fix if more than one app supports ACTION_OPEN_DOCUMENT/_CREATE_DOCUMENT
Android versions between (inclusive) 19 (4.4) and 29 (10) will show a
system intent target chooser if more than one app supports
ACTION_OPEN_DOCUMENT or ACTION_CREATE_DOCUMENT however the chooser will
not actually return the result of the running the app. This is
indistinguishable from not selecting a file at all.

This PR adds an app chooser that will be displayed if the situation is
detected.