Ensure display of multiple occurrences of the same element
If none or the same role for a relation member was set, it would only be
shown once in the relation membership tab even when occurring multiple
times. This was due to the relevant RelationMembers being equal to each
other (even though being different objects) and therefore List.indexOf
always returning the index of the 1st one in the List.