Correctly handle expanding/shrinking the offset array
This adds code to handle expanding/shrinking the offset array properly
when min or/and max zoom level is changed.

Previously this had only worked for increasing max zoom, but since the
introduction of the UI to configure custom imagery sources, this can
happen on the fly for both min and max zoom levels.