Support calling external measurement apps for appropriate preset fields
If a PresetTextField has one of ValueType.DIMENSION_HORIZONTAL or
ValueType.DIMENSION_VERTICAL set and a supported external app is
installed (currently only StreetMeasure), we will try and use it to set
the value.

For now we don't attempt to automatically or manually attempt to install
such apps as the consequences of that need to be though through.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2010