Check for Layer.NONE before retrieving tiles
This fixes the regression that we would try to retrieve tiles for the
NONE background layer (overlay layer was fine) and show a notification
for an incorrect URL.

The layer is no longer needed since we moved to the current layer
orchestration model, but is retained for people upgrading their installs
for now.

Includes some minor clean ups.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2441