Assure that UndoStorage is emptied correctly after an upload
This makes sure that the UndoStorage is reset before saving the current
state to ensure that it is not reinstated after restart and handles
removing undo data for selective uploads. Fixes a further issue in which
the failure listener was always called after an upload, as this is not
used in normal ops, this was a test only issue.