Import MSF files on Android 10+ and correctness fixes
This adds importing MSF files on Android 10 and later, including
displaying a tip the 1st time a file is "imported".

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2455

This further corrects the behaviour when files are selected and the
initial activity is paused and potentially destroyed by using the
restored activity and using the fragment manager to find fragments that
need to be updated from the imported data.