Use separate action section per element in OSC files
Creating elements that reference ones that are being undeleted would
previously fail, by using per element type action sections this can be
avoided except for relations, that is a newly created relation
referencing one that is going to be undeleted. Currently the later case
needs to be handled by first uploading the undeleted relations.