Respect multi-fetch limit when fetching elements for conflict resolution
Multi-fetch API call URL lengths can easily exceed the maximum URL
length. We currently limit the number of elements to 200 (same value as
JOSM uses) and simply loop to process all ids if more
are required.

Note: this unifies and simplifies the code as this was already "fixed" for certain situations.