Add the ISRG X1 certificate to the trust store used by OkHttp
This allows connections to sites using letsencrypt certificates to
continue to work for now on pre Android 7.1 devices. In particular this
affects the OSM API.

It is likely that this fix increases memory usage by multiple MBs.

Note: this does not solve the issue for things that do not use OkHttp,
for example ACRA.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/2556

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/1277