Bump API target in build script
Effects of this should largely be limited to image selection according
to https://developer.android.com/about/versions/14/changes/partial-photo-video-access
However, as normal, google contradicts itself in the document and is
very unclear if the compatible mode it supposedly uses will go away at
some time in the future and if it is any different to the previous
behaviour. In any case at least on Android 14 emulators there currently
doesn't seem to be any impact.

Resolves: https://github.com/MarcusWolschon/osmeditor4android/issues/2595