Add support for Link elements in Waypoints
This adds support for parsing Link elements in Waypoints and displaying
them as a clickable element in the Waypoint modal.

Note that the later is extremely painful due to SAF and having access
files relative to the actual GPX file. There's no guarantee that the
hack we've used here works reliably.

Resolves: https://github.com/MarcusWolschon/osmeditor4android/issues/2708