Improve behaviour and caching for large sequences
Panoramax has sequences in the order of magnitude of 10'000 entries
these take a substantial while to download, and we can't pass the whole
list to the viewer as that will exceed Androids max transaction buffer
size. We fix this by only passing roughly 200 ids to the viewer.

Further we limit the size of the sequence cache both for Panoramax and
Mapillary.