Automatically add an extension to files when saving if it is missing
THe Android system file picker does not allow to provide a default
extension to add to a file when creating a new one. Supposedly it should
do this for mime type to extension mappings it knows about, however that
doesn't seem to work either, and wouldn't solve the issue in any case as
it isn't possible to register additional mappings.

We get around this, at the price of this being a bad UX, by renaming the
file as soon as we have written it. WE only do this if there is no
extension at all present.

Further we attempt to save with proper mime types, however currently it
seems as if these are ignored by the system.

Note: currently we do nothing when saving scripts from the console
modal.

Resolves: https://github.com/MarcusWolschon/osmeditor4android/issues/2589