Save edit state immediately after (geo and josm) intent is processed
On some devices onPause and onResume may be called before the intent
processing in the initial onResume is finished, overwriting the editing
state before it can be saved. This change explicitly saves the editing
state immediately after the new viewbox/location has been set, and
synchronizes the corresponding code. This could potentially cause an ANR
in onResume but there doesn't seem to be a way around that.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/2833