Re-factor applicationId and app name handling
This decouples the applicationId from the Java package name and moves
app name and version configuration to the gradle build file, allowing:

- builds with different applicationIDs that can be installed (and run)
in parallel (for example for testing)
- all relevant configuration is in one place (Android gradle plugin
config)
- potential future renaming of the Java package.

The build with applicationId de.blau.android (used for googles
playstore) should be near identical to the previous build, the changes
add a build with the applicationId io.vespucci and app name vespucci.io.

The public directory used by the app is now derived from the app name,
this implies that that actually needs to be different for different
applicationIds as otherwise there might be access issues.