Use MediaStore.createDeleteRequest to delete images
On Android 11 and later we can ask the MEdiaStore to delete images this
works if if we are not owner of the image.

The down side of this is that depending on if the image is being indexed
via the media store or directly in the Vespucci Pictures directory, the
confirmation modal will be different.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/1913