Explicitly filter elements for clicking
Previously we relied on the cache of filtered elements being filled
prior to determining which objects can be clicked. This seems to fail
consistently since we switched to using SurfaceView in V20.

This switches to explicitly filtering objects, which resolves the issue.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2861