Ensure that we have enough space to add a further view in WrappingLayout
This fixes a ~13 year old bug when if the space remaining in a row of
preset views was one horizontal spacer or less a preset view was added
that ended up not being visible. This seems to have only been triggered
if accessibility screen zoom functions were used.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2960