Anchor popup menu on parent view if possible
If the Dialog used for the conflict resolution modal switches to
horizontal arrangement of the buttons, using the button as the anchor of
the popup menu can lead to it being cut off on the right hand side.

This PR uses the parent view of the button as the anchor if it exists
which will improve the behaviour somewhat, but naturally can't work
miracles: if the screen is not wide enough for the font size and text in
use nothing is going to fix the issue.

Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/2981