Force caching of way node filter results
If the filter result for a specific node is not cached, filtering in
getClickableNodes would return the initial result and not that after any
parent ways had been processed. By running filtering on any parent ways
we force the correct values to be cached.

Note that this does not do the same for relations which probably should
be done similarly.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/2997