Save list of images in file and not in bundle
Some, obviously very long, sequences were causing a TTLE this saves them
in a file and further doesn't save everything twice. This seems to be a
bug in Bundle as according to the doc adding a key a second time should
replace it and its value, but the bundle actually increases in size.