When merging nodes assure that the result is at the target node position
When merging nodes the source and target nodes may be swapped to
maintain as much history as possible. This change assures that the
resulting node is always at the target node position.

Note: this does not change the "merge into" semantics, aka the selected
node is merged in to the target node. This keeps the behavior consistent
with what happens when merging a node in to a way.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/3011