Workaround AGP 8 / Android 15 and later bug with per app languages
This turns off automatic generation of the locale list during the build
process that that causes Android 15 and later to hickup. This disables
the app language setting in the system preferences, adding the reference
to the locale config in the manifest seems to cause the issue which is
what the build option triggered.

Additionally improves activity restart behaviour when locale changes and
issue with language ordering in the preferences.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/3040

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/3041